Real locking starts to work for GBM, DRF2, GLM1, GLM2
Lockable.update takes the Job key doing the update.
SharedTreeBuilder locks & unlocks the input datasets.
DGLM locks & unlocks the xval models.

+++ src/test/java/hex/gbm/GBMTest.java
@@ -216,7 +216,7 @@ public class GBMTest extends TestUtil {
       preds.delete();
 
     } finally {
-      gbmmodel.delete();        // Remove the model
+      if( gbmmodel != null ) gbmmodel.delete(); // Remove the model
       gbm.source.delete();      // Remove original hex frame key
       UKV.remove(gbm.response._key);
       gbm.remove();             // Remove GBM Job
